dear cesium maintainer
Hi I am Rolf Aglaster (find me on the internet) and have startet translating the language file to a german version. We have a group of > 40 people now in Aachen and many of them are not used to reading french or not even Englisch. If it will be successful I can start over with duniter, gchange etc. as far support is needed. cheers Rolf